+++ js/test/qrcode.test.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" >
     <title>qrcode | default</title>
-    <script src="../qrcode.js"></script>
+    <script src="../dist/qrcode.js"></script>
     <script type="module">
 
 import { runTests } from '@web/test-runner-mocha';
