+++ js/test/qrcode.test.mjs
@@ -1,5 +1,5 @@
 
-import { qrcode } from '../qrcode.mjs';
+import { qrcode } from '../dist/qrcode.mjs';
 import { overview as test } from './qrcode-test-impl.js';
 
 test(qrcode);
