+++ js/test/qrcode_SJIS.test.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" >
     <title>qrcode | SJIS</title>
-    <script src="../qrcode.js"></script>
-    <script src="../qrcode_SJIS.js"></script>
+    <script src="../dist/qrcode.js"></script>
+    <script src="../dist/qrcode_SJIS.js"></script>
     <script type="module">
 
 import { runTests } from '@web/test-runner-mocha';
