+++ js/test/qrcode_SJIS.test.mjs
@@ -1,6 +1,6 @@
 
-import { qrcode } from '../qrcode.mjs';
-import { stringToBytes } from '../qrcode_SJIS.mjs';
+import { qrcode } from '../dist/qrcode.mjs';
+import { stringToBytes } from '../dist/qrcode_SJIS.mjs';
 import { sjis as test } from './qrcode-test-impl.js';
 import { misc } from './qrcode-test-impl.js';
