+++ js/test/qrcode_UTF8.test.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" >
     <title>qrcode | UTF8</title>
-    <script src="../qrcode.js"></script>
-    <script src="../qrcode_UTF8.js"></script>
+    <script src="../dist/qrcode.js"></script>
+    <script src="../dist/qrcode_UTF8.js"></script>
     <script type="module">
 
 import { runTests } from '@web/test-runner-mocha';
