+++ js/test/qrcode_UTF8.test.mjs
@@ -1,6 +1,6 @@
 
-import { qrcode } from '../qrcode.mjs';
-import { stringToBytes } from '../qrcode_UTF8.mjs';
+import { qrcode } from '../dist/qrcode.mjs';
+import { stringToBytes } from '../dist/qrcode_UTF8.mjs';
 import { utf8 as test } from './qrcode-test-impl.js';
 
 qrcode.stringToBytes = stringToBytes;
